input started
